requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
pillow
qrcode
groq